Add Options class and simplecov
  * Moved some search options to a dedicated class
  * Added some more testing
  * Added simplecov
