+++ .gitignore
@@ -11,3 +11,4 @@
 # rspec failure tracking
 Gemfile.lock
 .rspec_status
+.DS_Store
