+++ guardian_searcher.gemspec
@@ -34,6 +34,7 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'guard', "~> 2.18"
   spec.add_development_dependency 'guard-bundler', "~> 3.0"
   spec.add_development_dependency 'guard-rspec', "~> 4.7"
+  spec.add_development_dependency 'simplecov', '~> 0.21'
   # For more information and examples about making a new gem, checkout our
   # guide at: https://bundler.io/guides/creating_gem.html
 end
