+++ lib/guardian_searcher/base.rb
@@ -3,6 +3,7 @@
 module GuardianSearcher
   class Base
     include Faraday
+
     attr_reader :api_key
     attr_accessor :base_uri
 
@@ -31,27 +32,7 @@ module GuardianSearcher
     private
 
     def build_options(options)
-      return {} if options.empty?
-
-      opt = ""
-      options.each do |key, value|
-        # TODO: log a proper message if a key is not supported but continue with the ops building
-        next unless valid_option?(key)
-        opt += "&#{map_option(key)}=#{value}"
-      end
-    end
-
-    def valid_option?(option)
-      %i[from_date to_date page_size page].include?(option)
-    end
-
-    def map_option(key)
-      {
-        from_date: "from-date",
-        to_date: "to-date",
-        page_size: "page-size",
-        page: "page"
-      }[key]
+      Options.new(options).build_options
     end
   end
 end
