+++ lib/guardian_searcher/options.rb
@@ -14,5 +14,28 @@ module GuardianSearcher
       raise OptionsNotHashError unless options.is_a?(Hash)
       @options = options
     end
+
+    def build_options
+      return {} if options.empty?
+
+      opt = ""
+      options.each do |key, value|
+        valid_option?(key)
+        opt += "&#{map_option(key)}=#{value}"
+      end
+    end
+
+    def valid_option?(option)
+      raise OptionsNotSupportedError unless %i[from_date to_date page_size page].include?(option)
+    end
+
+    def map_option(key)
+      {
+        from_date: "from-date",
+        to_date: "to-date",
+        page_size: "page-size",
+        page: "page"
+      }[key]
+    end
   end
 end
